tests/contract/requirements.txt
pytest==8.3.4
pact-python==2.2.0
requests==2.32.3
